refactor(plans): extract plan settings drawer hook
Replace the imperative ref wrapper around useFormDrawer with a
usePlanSettingsDrawer hook returning { openDrawer }, matching the
subscription info drawer. The section calls the hook; cascade and
sub-mode plan-override logic stay encapsulated. Port the drawer test
to a hook test and switch the parent mocks to the hook.
